PAE-1675 removes basic auth from some endpoints
Removes basic auth as a strategy on 3x endpoints that are no longer used by the service providing reapply for accreditation

Also moves tests that exercise configured (or not) basic auth credentials out of integration tests for specific endpoints and into basic-auth-plugin.test.js
