+++ src/common/hapi-types.js
@@ -128,6 +128,7 @@
  * @property {Function} decorate - Decorate server/request with additional properties
  * @property {Function} dependency - Declare plugin dependencies
  * @property {Function} register - Register a plugin
+ * @property {Function} route - Register a route
  * @property {Function} start - Start the server
  * @property {Function} stop - Stop the server
  * @property {Function} initialize - Initialize the server without starting
