+++ src/overseas-sites/routes/get-by-id.js
@@ -5,8 +5,8 @@ import {
   LOGGING_EVENT_ACTIONS,
   LOGGING_EVENT_CATEGORIES
 } from '#common/enums/index.js'
+import { getAuthConfig } from '#common/helpers/auth/get-auth-config.js'
 import { SCOPES } from '#common/helpers/auth/constants.js'
-import { STRATEGY_NAME as BASIC_AUTH } from '#plugins/auth/basic-auth-plugin.js'
 
 /** @import { OverseasSitesRepository } from '#overseas-sites/repository/port.js' */
 
@@ -16,10 +16,7 @@ export const overseasSiteById = {
   method: 'GET',
   path: overseasSiteByIdPath,
   options: {
-    auth: {
-      strategies: ['access-token', BASIC_AUTH],
-      scope: [SCOPES.adminRead, SCOPES.organisationRead]
-    },
+    auth: getAuthConfig([SCOPES.adminRead]),
     tags: ['api']
   },
   /**
