+++ src/overseas-sites/routes/get-by-id.test.js
@@ -128,69 +128,4 @@ describe(`${overseasSiteByIdPath} route`, () => {
       })
     })
   })
-
-  describe('basic-auth strategy', () => {
-    const validCredentials = Buffer.from('basic-auth-user:changeme').toString(
-      'base64'
-    )
-
-    const buildBasicAuthServer = (basicAuth, repository) =>
-      createTestServer({
-        config: { basicAuth },
-        repositories: { overseasSitesRepository: () => repository },
-        featureFlags: createInMemoryFeatureFlags({})
-      })
-
-    it('returns 200 with valid Basic Auth credentials', async () => {
-      const repository = createInMemoryOverseasSitesRepository()()
-      const created = await repository.create(buildOverseasSite())
-      const server = await buildBasicAuthServer(
-        { username: 'basic-auth-user', password: 'changeme' },
-        repository
-      )
-
-      const response = await server.inject({
-        method: 'GET',
-        url: `/v1/overseas-sites/${created.id}`,
-        headers: { Authorization: `Basic ${validCredentials}` }
-      })
-
-      expect(response.statusCode).toBe(StatusCodes.OK)
-    })
-
-    it('returns 401 with wrong password', async () => {
-      const repository = createInMemoryOverseasSitesRepository()()
-      const created = await repository.create(buildOverseasSite())
-      const server = await buildBasicAuthServer(
-        { username: 'basic-auth-user', password: 'changeme' },
-        repository
-      )
-      const encoded = Buffer.from('basic-auth-user:wrong').toString('base64')
-
-      const response = await server.inject({
-        method: 'GET',
-        url: `/v1/overseas-sites/${created.id}`,
-        headers: { Authorization: `Basic ${encoded}` }
-      })
-
-      expect(response.statusCode).toBe(StatusCodes.UNAUTHORIZED)
-    })
-
-    it('returns 401 with valid credentials when basic-auth is not configured', async () => {
-      const repository = createInMemoryOverseasSitesRepository()()
-      const created = await repository.create(buildOverseasSite())
-      const server = await buildBasicAuthServer(
-        { username: '', password: '' },
-        repository
-      )
-
-      const response = await server.inject({
-        method: 'GET',
-        url: `/v1/overseas-sites/${created.id}`,
-        headers: { Authorization: `Basic ${validCredentials}` }
-      })
-
-      expect(response.statusCode).toBe(StatusCodes.UNAUTHORIZED)
-    })
-  })
 })
