+++ src/routes/v1/organisations/get.test.js
@@ -273,120 +273,6 @@ describe('GET /v1/organisations', () => {
     })
   })
 
-  describe('basic-auth strategy', () => {
-    const validCredentials = Buffer.from('basic-auth-user:changeme').toString(
-      'base64'
-    )
-
-    describe('when basic-auth credentials are configured', () => {
-      /**
-       * @type {import('#test/create-test-server.js').TestServer}
-       */
-      let server
-
-      beforeEach(async () => {
-        server = await createTestServer({
-          config: {
-            basicAuth: {
-              username: 'basic-auth-user',
-              password: 'changeme'
-            }
-          },
-          repositories: {
-            organisationsRepository: createInMemoryOrganisationsRepository([])
-          },
-          featureFlags: createInMemoryFeatureFlags()
-        })
-      })
-
-      it('returns 200 with valid Basic Auth credentials', async () => {
-        const response = await server.inject({
-          method: 'GET',
-          url: '/v1/organisations',
-          headers: { Authorization: `Basic ${validCredentials}` }
-        })
-
-        expect(response.statusCode).toBe(StatusCodes.OK)
-      })
-
-      it('returns 401 with wrong password', async () => {
-        const encoded = Buffer.from('basic-auth-user:wrong').toString('base64')
-        const response = await server.inject({
-          method: 'GET',
-          url: '/v1/organisations',
-          headers: { Authorization: `Basic ${encoded}` }
-        })
-
-        expect(response.statusCode).toBe(StatusCodes.UNAUTHORIZED)
-      })
-
-      it('returns 401 with wrong username', async () => {
-        const encoded = Buffer.from('wrong:changeme').toString('base64')
-        const response = await server.inject({
-          method: 'GET',
-          url: '/v1/organisations',
-          headers: { Authorization: `Basic ${encoded}` }
-        })
-
-        expect(response.statusCode).toBe(StatusCodes.UNAUTHORIZED)
-      })
-
-      it('returns 401 with malformed Basic Auth value', async () => {
-        const response = await server.inject({
-          method: 'GET',
-          url: '/v1/organisations',
-          headers: { Authorization: 'Basic notbase64credentials' }
-        })
-
-        expect(response.statusCode).toBe(StatusCodes.UNAUTHORIZED)
-      })
-
-      it('returns 401 with Authorization header that does not provide a basic auth value', async () => {
-        const response = await server.inject({
-          method: 'GET',
-          url: '/v1/organisations',
-          headers: { Authorization: 'not a basic auth value' }
-        })
-
-        expect(response.statusCode).toBe(StatusCodes.UNAUTHORIZED)
-      })
-
-      it('returns 401 with no Authorization header', async () => {
-        const response = await server.inject({
-          method: 'GET',
-          url: '/v1/organisations'
-        })
-
-        expect(response.statusCode).toBe(StatusCodes.UNAUTHORIZED)
-      })
-    })
-
-    describe('when basic-auth credentials are not configured', () => {
-      it('returns 401 with valid Basic Auth credentials', async () => {
-        const server = await createTestServer({
-          config: {
-            basicAuth: {
-              username: '', // matches default value in config.js
-              password: '' // matches default value in config.js
-            }
-          },
-          repositories: {
-            organisationsRepository: createInMemoryOrganisationsRepository([])
-          },
-          featureFlags: createInMemoryFeatureFlags()
-        })
-
-        const response = await server.inject({
-          method: 'GET',
-          url: '/v1/organisations',
-          headers: { Authorization: `Basic ${validCredentials}` }
-        })
-
-        expect(response.statusCode).toBe(StatusCodes.UNAUTHORIZED)
-      })
-    })
-  })
-
   testInvalidTokenScenarios({
     server: () => server,
     makeRequest: async () => {
